Mostly usable update? More testing needed! - Thread-safe Queue for SenderThread - Fix a few CSS glitches on macOS (wip), as well as fake   NSSegmentedButtons for the ChannelListWidget. - MatrixWidget is working...somewhat... - File association and icon dropping on macOS - Decrease Clang warnings with Q_NULLPTR - Use enums and constants when possible instead of using preprocessor   defines. - Removed some unused code - Sheet dialog tests on macOS. Haven't fully implemented it. - Copyright headers - I don't really know who and when to put on those. TODO:     - Easy way to switch channels and stuff since you can't scroll       QComboBoxes on macOS.     - More sheets on macOS     - File icons     - MiscWidget is still flickering.     - Retina/vector icons.     - Even more efficiency in repainting...still.         - Don't use QPixmap at all?         - Render in chunks?     - Use type() more often instead of qobject_cast?     - Instead of hacking with CSS, use native Cocoa.     - PianoWidget is still broken.     - Better handling of file dropping and command line args         - Multi-instance/tabs? Multi-instance doesn't work on macOS,           so now would be a great time to rework.     - Clean up extra comments and stuff.

+++ src/tool/SizeChangeTool.h
@@ -43,7 +43,7 @@ class SizeChangeTool : public EventTool {
 
 	private:
 		bool inDrag, dragsOnEvent;
-		int xPos;
+		qreal xPos;
 };
 
 #endif
